% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ranking.R
\name{calculate_dominance_ranking}
\alias{calculate_dominance_ranking}
\title{Theory of Dominance for Ranking Consensus}
\usage{
calculate_dominance_ranking(rank_mat)
}
\arguments{
\item{rank_mat}{A matrix where columns are methods and rows are alternatives.}
}
\value{
A numeric vector containing the final consensus ranking.
}
\description{
Calculates a consensus ranking based on the Theory of Dominance. It compares
multiple rankings and determines a final position based on majority rule
and dominance logic.
}
\keyword{internal}
