% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwm_solver.R
\name{.combine_constraints}
\alias{.combine_constraints}
\title{Internal Constraint Combiner}
\usage{
.combine_constraints(constraints, new_constraint)
}
\description{
Internal Constraint Combiner
}
\keyword{internal}
