% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwm_solver.R
\name{.validate_bwm_data}
\alias{.validate_bwm_data}
\title{Internal Data Validator}
\usage{
.validate_bwm_data(best_to_others, others_to_worst, criteria_names)
}
\description{
Prepares and validates comparison vectors.
}
\keyword{internal}
