% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.fuzzy_multimoora_res}
\alias{plot.fuzzy_multimoora_res}
\title{Plot MULTIMOORA Strategic Position Map}
\usage{
\method{plot}{fuzzy_multimoora_res}(x, ...)
}
\arguments{
\item{x}{An object of class `fuzzy_multimoora_res`.}

\item{...}{Additional arguments passed to plot.}
}
\description{
Visualizes the trade-off between the Ratio System (Additive Utility)
and the Reference Point (Min-Max Regret).
}
