% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.fuzzy_promethee_res}
\alias{plot.fuzzy_promethee_res}
\title{Plot PROMETHEE II Ranking}
\usage{
\method{plot}{fuzzy_promethee_res}(x, ...)
}
\arguments{
\item{x}{An object of class `fuzzy_promethee_res`.}

\item{...}{Additional arguments.}
}
\description{
Visualizes the Net Outranking Flows (Phi Net).
}
