% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.fuzzy_topsis_res}
\alias{plot.fuzzy_topsis_res}
\title{Plot TOPSIS Efficiency Map (Advanced)}
\usage{
\method{plot}{fuzzy_topsis_res}(x, ...)
}
\description{
Uses Distance Ratios for quadrants and connection lines.

NOW INCLUDES: Geometric Distance on dotted lines (Lower is Better).
}
