% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.fuzzy_vikor_res}
\alias{plot.fuzzy_vikor_res}
\title{Plot VIKOR Strategic Map (Advanced)}
\usage{
\method{plot}{fuzzy_vikor_res}(x, ...)
}
\description{
Uses Exponential Scaling for bubbles and Median Quadrants.
}
