% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.fuzzy_waspas_res}
\alias{plot.fuzzy_waspas_res}
\title{Plot WASPAS Consistency Map (Advanced)}
\usage{
\method{plot}{fuzzy_waspas_res}(x, ...)
}
\description{
Plot WASPAS Consistency Map (Advanced)
}
