% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ranking.R
\name{calculate_dominance_ranking}
\alias{calculate_dominance_ranking}
\title{Theory of Dominance for Ranking Consensus}
\usage{
calculate_dominance_ranking(r1, r2, r3)
}
\arguments{
\item{r1}{Numeric vector. First ranking.}

\item{r2}{Numeric vector. Second ranking.}

\item{r3}{Numeric vector. Third ranking.}
}
\value{
A numeric vector containing the final consensus ranking.
}
\description{
Calculates a consensus ranking based on the Theory of Dominance. It compares
three distinct rankings and determines a final position based on majority rule
and dominance logic.
}
\keyword{internal}
